FlaskBlog _Final/flaskblog/requirements.txt.txt
appdirs==1.4.4
asgiref==3.3.1
bcrypt==3.2.0
blinker==1.4
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
distlib==0.3.1
Django==3.1.5
dnspython==2.1.0
email-validator==1.1.2
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==8.1.0
pipenv==2020.11.15
pycparser==2.20
pytz==2020.5
requests==2.25.1
shell==1.0.1
six==1.15.0
SQLAlchemy==1.3.22
sqlparse==0.4.1
svgwrite==1.4.1
Tree==0.2.4
urllib3==1.26.2
virtualenv==20.3.1
virtualenv-clone==0.5.4
Werkzeug==1.0.1
WTForms==2.3.3